Moves Footer conditional display logic to MiniCart
