+++ packages/venia-concept/src/components/MiniCart/__tests__/footer.spec.js
@@ -14,44 +14,8 @@ const baseProps = {
     isMiniCartMaskOpen: false
 };
 
-test('renders totals summary and checkout components', () => {
+test('renders the correct tree', () => {
     const tree = renderer.render(<Footer {...baseProps} />);
-
-    expect(tree).toMatchSnapshot();
-});
-
-test('returns null when cart is empty', () => {
-    const props = {
-        ...baseProps,
-        isCartEmpty: true
-    };
-    const tree = renderer.render(<Footer {...props} />);
-
-    expect(tree).toMatchSnapshot();
-});
-
-test('returns null when cart is loading', () => {
-    const props = {
-        ...baseProps,
-        cart: {
-            ...baseProps.cart,
-            isLoading: true
-        }
-    };
-    const tree = renderer.render(<Footer {...props} />);
-
-    expect(tree).toMatchSnapshot();
-});
-
-test('returns null when cart is editing item', () => {
-    const props = {
-        ...baseProps,
-        cart: {
-            ...baseProps.cart,
-            isEditingItem: true
-        }
-    };
-    const tree = renderer.render(<Footer {...props} />);
 
     expect(tree).toMatchSnapshot();
 });
